% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSpecWavelet-package.R
\docType{data}
\name{exampleMS}
\alias{exampleMS}
\title{An example mass spectrum}
\format{
A numeric vector represents the mass spectrum with equal sample
intervals.
}
\source{
CAMDA, CAMDA 2006 Competition Data Set. 2006, http://camda.duke.edu.
}
\description{
An example mass spectrum from CAMDA 2006. All-in-1 Protein Standard II
(Ciphergen Cat. # C100-0007) were measured on Ciphergen NP20 chips. There
are 7 polypeptides in the sample with m/z values of 7034, 12230, 16951,
29023, 46671, 66433, 147300.
}
\keyword{datasets}
