% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendLength.R
\name{extendLength}
\alias{extendLength}
\title{Extend the length of a signal or matrix}
\usage{
extendLength(
  x,
  addLength = NULL,
  method = c("reflection", "open", "circular"),
  direction = c("right", "left", "both")
)
}
\arguments{
\item{x}{a vector or matrix with column with each column as a signal}

\item{addLength}{the length to be extended}

\item{method}{three methods available, c("reflection", "open", "circular").
By default, it is "reflection".}

\item{direction}{three options available: c("right", "left", "both")}
}
\value{
return the extended vector or matrix.
}
\description{
Extend the length of a signal or matrix by row
}
\examples{

# a = matrix(rnorm(9), 3)
# extendLength(a, 3, direction='right') 	## not exposed function

}
\seealso{
\code{\link[=extendNBase]{extendNBase()}}
}
\author{
Pan Du
}
\keyword{methods}
