% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMaximum.R
\name{localMaximum}
\alias{localMaximum}
\title{Identify local maximum within a slide window.}
\usage{
localMaximum(x, winSize = 5)
}
\arguments{
\item{x}{a vector represents a signal profile}

\item{winSize}{the slide window size, 5 by default.}
}
\value{
Return a vector with the same length of the input x. The position of
local maximum is set as 1L, 0L else where.
}
\description{
Identify local maximum within a slide window.
}
\examples{

x <- rnorm(200)
lmax <- localMaximum(x, 5)
maxInd <- which(lmax > 0)
plot(x, type = "l")
points(maxInd, x[maxInd], col = "red")

}
\seealso{
\code{\link[=getLocalMaximumCWT]{getLocalMaximumCWT()}}
}
\author{
Pan Du and Sergio Oller
}
\keyword{methods}
