% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexh.R
\name{mexh}
\alias{mexh}
\title{The mexican hat function}
\usage{
mexh(x)
}
\arguments{
\item{x}{where to evaluate the mexican hat}
}
\value{
A vector of the same length as \code{x} with the corresponding values
}
\description{
\deqn{ \psi(x) = \frac{2}{\sqrt{3}} \pi^{-0.25} ( 1 - x^2 ) \exp{-x^2/2} }
}
\examples{
x <- seq(-8, 8, length.out = 256)
mexh(x)
}
