% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzInd2vRange.R
\name{mzInd2vRange}
\alias{mzInd2vRange}
\title{Match m/z index to m/z value with a certain error range}
\usage{
mzInd2vRange(mzInd, error = 0.003)
}
\arguments{
\item{mzInd}{a vector of m/z index}

\item{error}{error range}
}
\value{
return a vector of sorted m/z values
}
\description{
Match m/z index to m/z value with a certain error range
}
\details{
\strong{Deprecated}: this function appears to be unused, both internally in
MassSpecWavelet and by known downstream packages, and is scheduled for
removal in a future release. If you rely on it, please open an issue at
\url{https://github.com/zeehio/MassSpecWavelet/issues} within the next year.
}
\seealso{
\code{\link[=mzV2indRange]{mzV2indRange()}}
}
\author{
Pan Du
}
\keyword{methods}
