% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzV2indRange.R
\name{mzV2indRange}
\alias{mzV2indRange}
\title{Match m/z value to m/z index with a certain error range}
\usage{
mzV2indRange(mzV, error = 0.003)
}
\arguments{
\item{mzV}{a vector of m/z value}

\item{error}{error range}
}
\value{
return a vector of sorted m/z indexes
}
\description{
Match m/z value to m/z index with a certain error range
}
\details{
\strong{Deprecated}: this function appears to be unused, both internally in
MassSpecWavelet and by known downstream packages, and is scheduled for
removal in a future release. If you rely on it, please open an issue at
\url{https://github.com/zeehio/MassSpecWavelet/issues} within the next year.
}
\seealso{
\code{\link[=mzInd2vRange]{mzInd2vRange()}}
}
\author{
Pan Du
}
\keyword{methods}
