% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocalMaximumCWT.R
\name{getLocalMaximumCWT}
\alias{getLocalMaximumCWT}
\title{Identify the local maximum of each column in 2-D CWT coefficients matrix}
\usage{
getLocalMaximumCWT(
  wCoefs,
  minWinSize = 5,
  amp.Th = 0,
  is_amp_thres_relative = FALSE
)
}
\arguments{
\item{wCoefs}{2-D CWT coefficients, each column corresponding to CWT
coefficient at one scale. The column name is the scale.}

\item{minWinSize}{The minimum slide window size used.}

\item{amp.Th}{The minimum peak amplitude.}

\item{is_amp_thres_relative}{Whether \code{amp.Th} is given relative to
\code{max(wCoefs)}.}
}
\value{
return a matrix with same dimension as CWT coefficient matrix,
wCoefs. The local maxima are marked as 1, others are 0.
}
\description{
Identify the local maximum of each column in 2-D CWT coefficients matrix by
using a slide window. The size of slide window linearly changes from the
coarse scale (bigger window size) to detail scale. The scale of CWT
increases with the column index.
}
\examples{

data(exampleMS)
scales <- seq(1, 64, 3)
wCoefs <- cwt(exampleMS[5000:11000], scales = scales, wavelet = "mexh")

localMax <- getLocalMaximumCWT(wCoefs)
plotLocalMax(localMax)

}
\seealso{
\code{\link[=localMaximum]{localMaximum()}}
}
\author{
Pan Du
}
\keyword{methods}
