% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRidgeValue.R
\name{getRidgeValue}
\alias{getRidgeValue}
\title{Get the CWT coefficient values corresponding to the peak ridge}
\usage{
getRidgeValue(ridgeList, wCoefs, skip = 0)
}
\arguments{
\item{ridgeList}{a list of ridge lines}

\item{wCoefs}{2-D CWT coefficients}

\item{skip}{the CWT scale level to be skipped, by default the 0 scale level
(raw spectrum) is skipped.}
}
\value{
A list of ridge values corresponding to the input ridgeList.
}
\description{
Get the CWT coefficient values corresponding to the peak ridge
}
\author{
Pan Du
}
\keyword{methods}
