% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMaximum.R
\name{localMaximum}
\alias{localMaximum}
\title{Identify local maximum within a slide window.}
\usage{
localMaximum(x, winSize = 5)
}
\arguments{
\item{x}{a vector represents a signal profile}

\item{winSize}{the slide window size, 5 by default.}
}
\value{
Return a vector with the same length of the input x. The position of
local maximum is set as 1, 0 else where.
}
\description{
Find local maximum by transform the vector as matrix, then get the the
maximum of each column. This operation is performed twice with vecctor
shifted half of the winSize.
}
\details{
Instead of find the local maximum by a slide window, which slide all
possible positions, we find local maximum by transform the vector as matrix,
then get the the maximum of each column. This operation is performed twice
with vecctor shifted half of the winSize. The main purpose of this is to
increase the efficiency of the algorithm.
}
\examples{

x <- rnorm(200)
lmax <- localMaximum(x, 5)
maxInd <- which(lmax > 0)
plot(x, type='l')
points(maxInd, x[maxInd], col='red')

}
\seealso{
\code{\link{getLocalMaximumCWT}}
}
\author{
Pan Du
}
\keyword{methods}
