% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzV2indRange.R
\name{mzV2indRange}
\alias{mzV2indRange}
\title{Match m/z value to m/z index with a certain error range}
\usage{
mzV2indRange(mzV, error = 0.003)
}
\arguments{
\item{mzV}{a vector of m/z value}

\item{error}{error range}
}
\value{
return a vector of sorted m/z indexes
}
\description{
Match m/z value to m/z index with a certain error range
}
\seealso{
\code{\link{mzInd2vRange}}
}
\author{
Pan Du
}
\keyword{methods}
